+++ src/App.css
@@ -1,6 +1,6 @@
 :root {
   --primary: #fff;
-  --pink: #347cf0e5;
-  --dark-pink: #1d83bedc;
-  --light-pink: #3b72e7d2;
+  --pink: #f03f83;
+  --dark-pink: #ec4c8a;
+  --light-pink: #fc5193;
 }
