+++ src/components/css/items.module.css
@@ -58,8 +58,6 @@
   width: 100%;
   max-width: 400px;
   opacity: 0.3;
-
-  /* filter: brightness(50%) grayscale(50%) opacity(0.5); */
 }
 
 .item-btns {
